apache: disallow access to all directories and files starting with a dot
Except for the .well-known directory there shouldn't be any cases where
these paths should be accessible, and in some cases they might actually
contain sensitive information.
The .well-known directory is used by Let's Encrypt and must be reachable,
but since it is an alias to a non-dot directory it still works.