php: don't install php5-imagick any more
This module is not currently installed on the webserver (which is the only
playbook that uses the php role directly or through dependencies). And I
don't know why it was added to this list in the first place.

Also, php5-imagick is not available on Debian Stretch anyway (only
php-imagick is available).